Move shared option resources to src root folders
