+++ src/components/covariant-menu/hooks/use-nudge-harp-strata-by-harp-key/use-nudge-harp-strata-by-harp-key.ts
@@ -7,7 +7,7 @@ import {
 } from 'harpstrata'
 import type { HarpStrata } from 'harpstrata'
 
-import { partiallyApplyNudgeFunction } from '../../../layout-menu/utils'
+import { partiallyApplyNudgeFunction } from '../../../../utils'
 import type { SetActiveHarpStrata } from '../../../../types'
 import { DisplayModes } from '../../../../types'
 import { getPropsForHarpStrata } from '../../../../helpers'
