+++ src/components/covariant-menu/hooks/use-nudge-harp-strata-by-pozition/use-nudge-harp-strata-by-pozition.ts
@@ -7,7 +7,7 @@ import {
 } from 'harpstrata'
 import type { HarpStrata } from 'harpstrata'
 
-import { partiallyApplyNudgeFunction } from '../../../layout-menu/utils'
+import { partiallyApplyNudgeFunction } from '../../../../utils'
 import { DisplayModes } from '../../../../types'
 import type { SetActiveHarpStrata } from '../../../../types'
 import { getPropsForHarpStrata } from '../../../../helpers'
