+++ src/components/layout-menu/hooks/index.ts
@@ -1,3 +1,2 @@
-export { useNudgeDisplayMode } from './use-nudge-display-mode'
 export { useNudgeExperienceMode } from './use-nudge-experience-mode'
 export { useNudgeHarpStrataByApparatus } from './use-nudge-harp-strata-by-apparatus'
