+++ src/components/layout-menu/hooks/use-nudge-harp-strata-by-apparatus/use-nudge-harp-strata-by-apparatus.ts
@@ -2,7 +2,7 @@ import { useGlobal } from 'reactn'
 import { getHarpStrata, getApparatusIds } from 'harpstrata'
 import type { HarpStrata, ApparatusIds } from 'harpstrata'
 
-import { partiallyApplyNudgeFunction } from '../../utils'
+import { partiallyApplyNudgeFunction } from '../../../../utils'
 import { DisplayModes } from '../../../../types'
 import type { SetActiveHarpStrata } from '../../../../types'
 import { getPropsForHarpStrata } from '../../../../helpers'
