--- src/components/layout-menu/utils/index.ts
@@ -1,1 +1,0 @@
-export { partiallyApplyNudgeFunction } from './paritally-apply-nudge-function'
