+++ src/utils/index.ts
@@ -1,3 +1,4 @@
 export { setGlobalState } from './set-global-state'
 export { setGlobalReducers } from './set-global-reducers'
 export { getHarpFaceFacts } from './get-harp-face-facts'
+export { partiallyApplyNudgeFunction } from './paritally-apply-nudge-function'
